# deal/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8082
  description: Generated server url
paths:
  /deal/statement:
    post:
      tags:
      - deal-api-controller
      operationId: calculateLoanTerms
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanStatementRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LoanOfferDto"
  /deal/offer/select:
    post:
      tags:
      - deal-api-controller
      operationId: selectOffer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanOfferDto"
        required: true
      responses:
        "200":
          description: OK
  /deal/document/{statementId}/sign:
    post:
      tags:
      - deal-api-controller
      operationId: signDocumentRequest
      parameters:
      - name: statementId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
  /deal/document/{statementId}/send:
    post:
      tags:
      - deal-api-controller
      operationId: sendDocumentRequest
      parameters:
      - name: statementId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
  /deal/document/{statementId}/code:
    post:
      tags:
      - deal-api-controller
      operationId: signDocument
      parameters:
      - name: statementId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SesCodeDto"
        required: true
      responses:
        "200":
          description: OK
  /deal/calculate/{statementId}:
    post:
      tags:
      - deal-api-controller
      operationId: finishRegistrationAndCalculate
      parameters:
      - name: statementId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinishRegistrationRequestDto"
        required: true
      responses:
        "200":
          description: OK
components:
  schemas:
    LoanStatementRequestDto:
      type: object
      properties:
        amount:
          type: number
        term:
          type: integer
          format: int32
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        email:
          type: string
        birthdate:
          type: string
          format: date
        passportSeries:
          type: string
        passportNumber:
          type: string
    LoanOfferDto:
      type: object
      properties:
        statementId:
          type: string
          format: uuid
        requestedAmount:
          type: number
        totalAmount:
          type: number
        term:
          type: integer
          format: int32
        monthlyPayment:
          type: number
        rate:
          type: number
        isInsuranceEnabled:
          type: boolean
        isSalaryClient:
          type: boolean
    SesCodeDto:
      type: object
      properties:
        sesCode:
          type: integer
          format: int32
      required:
      - sesCode
    EmploymentDto:
      type: object
      properties:
        employmentStatus:
          type: string
          enum:
          - BUSINESS_OWNER
          - EMPLOYED
          - SELF_EMPLOYED
          - UNEMPLOYED
        employerINN:
          type: string
        salary:
          type: number
        position:
          type: string
          enum:
          - TOP_MANAGEMENT
          - MID_MANAGEMENT
          - WORKER
          - OWNER
        workExperienceTotal:
          type: integer
          format: int32
        workExperienceCurrent:
          type: integer
          format: int32
    FinishRegistrationRequestDto:
      type: object
      properties:
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
          - NON_BINARY
        maritalStatus:
          type: string
          enum:
          - SINGLE
          - MARRIED
          - DIVORCED
          - WIDOW_WIDOWER
        dependentAmount:
          type: integer
          format: int32
        passportIssueDate:
          type: string
          format: date
        passportIssueBranch:
          type: string
        employment:
          $ref: "#/components/schemas/EmploymentDto"
        accountNumber:
          type: string